+++ dm/pie_collection_member.sql
@@ -1,4 +1,5 @@
 CREATE TABLE pie_collection_member(
-       cmb_col_id INTEGER NOT NULL,
+       cmb_col_id BIGINT NOT NULL,
        cmb_mob_id BIGINT NOT NULL,
        PRIMARY KEY(cmb_col_id, cmb_mob_id));
+
